Navbar language linking (#36)
* Fixed language linking issues

Fixed problem that the home button and logo always linked to the English home site. Now they link to the current language.

* Language Links link to current page

The language links in the header link to the currently open page in their corresponding language. Earlier they linked to the Home page of that language.
Additionally the weird spacing and slashes are fixed (Removed them ;p).